[fix]Notification Heads-Up

+++ app/src/main/java/freeprojects/oldbigbuddha/kyoto/alarmapplication/Receivers/AlarmReceiver.java
@@ -21,7 +21,6 @@ public class AlarmReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
-        Realm.init(context);
         NotificationManager manager = (NotificationManager)context.getSystemService(Context.NOTIFICATION_SERVICE);
         NotificationCompat.Builder builder = new NotificationCompat.Builder(context);
 
@@ -33,17 +32,21 @@ public class AlarmReceiver extends BroadcastReceiver {
                 .setContentTitle(title)
                 .setContentText(content)
                 .setContentInfo("Alarm")
+                .setPriority(Notification.PRIORITY_HIGH)
+                .setDefaults(Notification.DEFAULT_ALL)
+                .setVisibility(Notification.VISIBILITY_PRIVATE)
                 .setWhen(System.currentTimeMillis());
 
         manager.notify(0, builder.build());
 
         Realm realm = Realm.getDefaultInstance();
         AlarmRealmData deleteData = realm.where(AlarmRealmData.class).equalTo("geofenceId", id).findFirst();
-        realm.beginTransaction();
-        // FIXME: 2017/08/18 Error now
-        deleteData.deleteFromRealm();
-        realm.commitTransaction();
-
+        if (deleteData != null) {
+            realm.beginTransaction();
+            // FIXME: 2017/08/18 Error now
+            deleteData.deleteFromRealm();
+            realm.commitTransaction();
+        }
 
         Log.d("onReceive", "はいったお！！！！！！！！");
     }
